+++ tests/test_format_beacon.py
@@ -10,6 +10,6 @@ from wdmapper.link import Link
 
 def test_write_no_header():
     out = io.StringIO('')
-    writer = beacon.writer(out, header=False)
+    writer = beacon.Writer(out, header=False)
     writer.write_link(Link('foo', 'b"r'))
     assert out.getvalue() == 'foo||b"r\n'