+++ tests/test_format_csv.py
@@ -19,7 +19,7 @@ def test_read():
 
 def test_write():
     out = io.StringIO("")
-    writer = csv.writer(out, header=True)
+    writer = csv.Writer(out, header=True)
     writer.init({})
     assert out.getvalue() == "source, target, annotation\n"